Put view tests into separate files.
