+++ tests/test_enterprise/views/test_course_enrollment_view.py
@@ -1,6 +1,8 @@
+# -*- coding: utf-8 -*-
 """
-Module tests user-facing views of the Enterprise app.
+Tests for the ``CourseEnrollmentView`` view of the Enterprise app.
 """
+
 from __future__ import absolute_import, unicode_literals
 
 import ddt
@@ -11,14 +13,12 @@ from pytest import mark
 
 from django.conf import settings
 from django.core.urlresolvers import reverse
-from django.shortcuts import render
 from django.test import Client, TestCase
 
 from enterprise.models import EnterpriseCourseEnrollment, EnterpriseCustomerUser
-from enterprise.utils import NotConnectedToOpenEdX
-from enterprise.views import LMS_COURSEWARE_URL, LMS_DASHBOARD_URL, LMS_START_PREMIUM_COURSE_FLOW_URL, HttpClientError
-# pylint: disable=import-error,wrong-import-order
-from six.moves.urllib.parse import urlencode
+from enterprise.views import HttpClientError
+from six.moves.urllib.parse import urlencode  # pylint: disable=import-error
+from test_utils import fake_render
 from test_utils.factories import (
     DataSharingConsentFactory,
     EnterpriseCourseEnrollmentFactory,
@@ -29,730 +29,6 @@ from test_utils.factories import (
 )
 from test_utils.fake_catalog_api import FAKE_COURSE, FAKE_COURSE_RUN, setup_course_catalog_api_client_mock
 from test_utils.mixins import MessagesMixin
-
-
-def fake_render(request, template, context):  # pylint: disable=unused-argument
-    """
-    Switch the request to use a template that does not depend on edx-platform.
-    """
-    return render(request, 'enterprise/emails/user_notification.html', context=context)
-
-
-@mark.django_db
-@ddt.ddt
-class TestProgramDataSharingPermissions(TestCase):
-    """
-    Test the user-facing program consent view
-    """
-
-    url = reverse('grant_data_sharing_permissions')
-
-    valid_get_params = {
-        'enterprise_customer_uuid': 'fake-uuid',
-        'next': 'https://google.com/',
-        'failure_url': 'https://facebook.com/',
-        'program_uuid': 'fake-program-uuid',
-    }
-    valid_post_params = {
-        'enterprise_customer_uuid': 'fake-uuid',
-        'redirect_url': 'https://google.com/',
-        'failure_url': 'https://facebook.com/',
-        'program_uuid': 'fake-program-uuid',
-        'data_sharing_consent': 'true',
-    }
-
-    def setUp(self):
-        self.user = UserFactory.create(username='john', is_staff=True, is_active=True)
-        self.user.set_password("QWERTY")
-        self.user.save()
-        self.client = Client()
-        config = mock.patch('enterprise.views.configuration_helpers')
-        self.configuration_helpers = config.start()
-        self.addCleanup(config.stop)
-        get_dsc = mock.patch('enterprise.views.get_data_sharing_consent')
-        self.get_data_sharing_consent = get_dsc.start()
-        self.addCleanup(get_dsc.stop)
-        super(TestProgramDataSharingPermissions, self).setUp()
-
-    def _login(self):
-        """
-        Log user in.
-        """
-        assert self.client.login(username=self.user.username, password="QWERTY")
-
-    def _mock_get_program_by_uuid(self):
-        """
-        Mock the calls needed to get a program from the Course Discovery API
-        """
-        discovery_client_class = mock.patch('enterprise.views.CourseCatalogApiServiceClient')
-        return_val = discovery_client_class.start().return_value.get_program_by_uuid
-        self.addCleanup(discovery_client_class.stop)
-        return return_val
-
-    @ddt.data(
-        'enterprise_customer_uuid',
-        'next',
-        'failure_url',
-        'program_uuid',
-    )
-    def test_get_program_consent_missing_parameter(self, missing_parameter):
-        self._mock_get_program_by_uuid().return_value = True
-        valid_get_params = self.valid_get_params.copy()
-        valid_get_params.pop(missing_parameter)
-        self._login()
-        response = self.client.get(self.url, valid_get_params)
-        assert response.status_code == 404
-
-    def test_get_consent_program_does_not_exist(self):
-        self._mock_get_program_by_uuid().side_effect = HttpClientError
-        self._login()
-        response = self.client.get(self.url, self.valid_get_params)
-        assert response.status_code == 404
-
-    def test_get_program_consent_no_ec(self):
-        self.get_data_sharing_consent.return_value = None
-        self._mock_get_program_by_uuid().return_value = True
-        self._login()
-        response = self.client.get(self.url, self.valid_get_params)
-        assert response.status_code == 404
-
-    def test_get_program_consent_not_required(self):
-        self.get_data_sharing_consent.return_value.consent_required.return_value = False
-        self._mock_get_program_by_uuid().return_value = True
-        self._login()
-        response = self.client.get(self.url, self.valid_get_params)
-        assert response.status_code == 404
-
-    @ddt.data(
-        'enterprise_customer_uuid',
-        'redirect_url',
-        'failure_url',
-        'program_uuid',
-    )
-    def test_post_program_consent_missing_parameter(self, missing_parameter):
-        self._mock_get_program_by_uuid().return_value = True
-        valid_post_params = self.valid_post_params.copy()
-        valid_post_params.pop(missing_parameter)
-        self._login()
-        response = self.client.post(self.url, valid_post_params)
-        assert response.status_code == 404
-
-    def test_post_consent_program_does_not_exist(self):
-        self._mock_get_program_by_uuid().side_effect = HttpClientError
-        self._login()
-        response = self.client.post(self.url, self.valid_post_params)
-        assert response.status_code == 404
-
-    def test_post_program_consent_no_ec(self):
-        self.get_data_sharing_consent.return_value = None
-        self._mock_get_program_by_uuid().return_value = True
-        self._login()
-        response = self.client.post(self.url, self.valid_post_params)
-        assert response.status_code == 404
-
-    def test_post_program_consent_not_required(self):
-        self.get_data_sharing_consent.return_value.consent_required.return_value = False
-        self._mock_get_program_by_uuid().return_value = True
-        self._login()
-        response = self.client.get(self.url, self.valid_post_params)
-        assert response.status_code == 404
-
-    def test_post_program_consent(self):
-        consent_record = self.get_data_sharing_consent.return_value
-        consent_record.consent_required.return_value = True
-        self._mock_get_program_by_uuid().return_value = True
-        self._login()
-        response = self.client.post(self.url, self.valid_post_params, follow=False)
-        consent_record.save.assert_called_once()
-        self.assertRedirects(response, 'https://google.com/', fetch_redirect_response=False)
-
-    def test_post_program_consent_not_provided(self):
-        consent_record = self.get_data_sharing_consent.return_value
-        consent_record.consent_required.return_value = True
-        self._mock_get_program_by_uuid().return_value = True
-        self._login()
-        params = self.valid_post_params.copy()
-        params.pop('data_sharing_consent')
-        response = self.client.post(self.url, params, follow=False)
-        consent_record.save.assert_called_once()
-        self.assertRedirects(response, 'https://facebook.com/', fetch_redirect_response=False)
-
-    def test_post_program_consent_deferred(self):
-        consent_record = self.get_data_sharing_consent.return_value
-        consent_record.consent_required.return_value = True
-        self._mock_get_program_by_uuid().return_value = True
-        self._login()
-        params = self.valid_post_params.copy()
-        params['enrollment_deferred'] = 'True'
-        response = self.client.post(self.url, params, follow=False)
-        consent_record.save.assert_not_called()
-        self.assertRedirects(response, 'https://google.com/', fetch_redirect_response=False)
-
-    @ddt.data(False, True)
-    @mock.patch('enterprise.views.render', side_effect=fake_render)
-    def test_get_program_consent(self, enrollment_deferred, mock_render):  # pylint: disable=unused-argument
-        self.get_data_sharing_consent.return_value.consent_required.return_value = True
-        enterprise_customer = self.get_data_sharing_consent.return_value.enterprise_customer
-        enterprise_customer.name = 'Starfleet Academy'
-        self._mock_get_program_by_uuid().return_value = True
-        self._login()
-        self.configuration_helpers.get_value.return_value = "My Platform"
-        params = self.valid_get_params.copy()
-        if enrollment_deferred:
-            params.update({'enrollment_deferred': True})
-        response = self.client.get(self.url, params)
-        assert response.status_code == 200
-
-        expected_prompt = (
-            'To access this program, you must first consent to share your learning achievements '
-            'with <b>Starfleet Academy</b>.'
-        )
-        expected_alert = (
-            'In order to start this program and use your discount, <b>you must</b> consent to share your '
-            'program data with Starfleet Academy.'
-        )
-
-        for key, value in {
-                "platform_name": "My Platform",
-                "consent_request_prompt": expected_prompt,
-                "requested_permissions_header": (
-                    'Per the <a href="#consent-policy-dropdown-bar" '
-                    'class="policy-dropdown-link background-input failure-link" id="policy-dropdown-link">'
-                    'Data Sharing Policy</a>, <b>Starfleet Academy</b> would like to know about:'
-                ),
-                'confirmation_alert_prompt': expected_alert,
-                'confirmation_alert_prompt_warning': '',
-                'sharable_items_footer': (
-                    'My permission applies only to data from courses or programs that are sponsored by '
-                    'Starfleet Academy, and not to data from any My Platform courses or programs that '
-                    'I take on my own. I understand that once I grant my permission to allow data to be shared '
-                    'with Starfleet Academy, I may not withdraw my permission but I may elect to unenroll '
-                    'from any courses that are sponsored by Starfleet Academy.'
-                ),
-                "program_uuid": params.get('program_uuid'),
-                "redirect_url": "https://google.com/",
-                "failure_url": "https://facebook.com/",
-                "enterprise_customer_name": enterprise_customer.name,
-                "program_specific": True,
-                "enrollment_deferred": enrollment_deferred,
-                "welcome_text": "Welcome to My Platform.",
-                'sharable_items_note_header': 'Please note',
-                "enterprise_customer": enterprise_customer
-        }.items():
-            assert response.context[key] == value  # pylint:disable=no-member
-
-
-@mark.django_db
-@ddt.ddt
-class TestGrantDataSharingPermissions(MessagesMixin, TestCase):
-    """
-    Test GrantDataSharingPermissions.
-    """
-
-    def setUp(self):
-        self.user = UserFactory.create(is_staff=True, is_active=True)
-        self.user.set_password("QWERTY")
-        self.user.save()
-        self.client = Client()
-        super(TestGrantDataSharingPermissions, self).setUp()
-
-    url = reverse('grant_data_sharing_permissions')
-
-    def _login(self):
-        """
-        Log user in.
-        """
-        assert self.client.login(username=self.user.username, password="QWERTY")
-
-    def _assert_enterprise_linking_messages(self, response, user_is_active=True):
-        """
-        Verify response messages for a learner when he/she is linked with an
-        enterprise depending on whether the learner has activated the linked
-        account.
-        """
-        response_messages = self._get_messages_from_response_cookies(response)
-        if user_is_active:
-            # Verify that request contains the expected success message when a
-            # learner with activated account is linked with an enterprise
-            self.assertEqual(len(response_messages), 1)
-            self._assert_request_message(
-                response_messages[0],
-                'success',
-                '<strong>Account created</strong> Thank you for creating an account with edX.'
-            )
-        else:
-            # Verify that request contains the expected success message and an
-            # info message when a learner with unactivated account is linked
-            # with an enterprise.
-            self.assertEqual(len(response_messages), 2)
-            self._assert_request_message(
-                response_messages[0],
-                'success',
-                '<strong>Account created</strong> Thank you for creating an account with edX.'
-            )
-            self._assert_request_message(
-                response_messages[1],
-                'info',
-                '<strong>Activate your account</strong> Check your inbox for an activation email. '
-                'You will not be able to log back into your account until you have activated it.'
-            )
-
-    def test_get_no_patches(self):
-        """
-        Test that we get the right exception when nothing is patched.
-        """
-        client = Client()
-        with self.assertRaises(NotConnectedToOpenEdX):
-            client.get(self.url)
-
-    @mock.patch('enterprise.views.render', side_effect=fake_render)
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.models.CourseCatalogApiServiceClient')
-    @mock.patch('enterprise.views.CourseApiClient')
-    @ddt.data(
-        (False, False, True),
-        (False, True, False),
-        (True, True, False),
-    )
-    @ddt.unpack
-    def test_get_course_specific_consent(
-            self,
-            enrollment_deferred,
-            supply_customer_uuid,
-            existing_course_enrollment,
-            course_api_client_mock,
-            course_catalog_api_client_mock,
-            mock_config,
-            *args
-    ):  # pylint: disable=unused-argument
-        course_id = 'course-v1:edX+DemoX+Demo_Course'
-        mock_config.get_value.return_value = 'My Platform'
-        course_catalog_api_client_mock.return_value.course_in_catalog.return_value = True
-        client = course_api_client_mock.return_value
-        client.get_course_details.return_value = {
-            'name': 'edX Demo Course',
-        }
-        self._login()
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-        )
-        ecu = EnterpriseCustomerUserFactory(
-            user_id=self.user.id,
-            enterprise_customer=enterprise_customer
-        )
-        if existing_course_enrollment:
-            EnterpriseCourseEnrollment.objects.create(
-                enterprise_customer_user=ecu,
-                course_id=course_id
-            )
-        params = {
-            'course_id': 'course-v1:edX+DemoX+Demo_Course',
-            'next': 'https://google.com'
-        }
-        if enrollment_deferred:
-            params['enrollment_deferred'] = True
-        if supply_customer_uuid:
-            params['enterprise_id'] = str(enterprise_customer.uuid)
-        response = self.client.get(self.url, data=params)
-        assert response.status_code == 200
-        expected_prompt = (
-            'To access this course, you must first consent to share your learning achievements '
-            'with <b>Starfleet Academy</b>.'
-        )
-        expected_alert = (
-            'In order to start this course and use your discount, <b>you must</b> consent to share your '
-            'course data with Starfleet Academy.'
-        )
-
-        for key, value in {
-                "platform_name": "My Platform",
-                "consent_request_prompt": expected_prompt,
-                "requested_permissions_header": (
-                    'Per the <a href="#consent-policy-dropdown-bar" '
-                    'class="policy-dropdown-link background-input failure-link" id="policy-dropdown-link">'
-                    'Data Sharing Policy</a>, <b>Starfleet Academy</b> would like to know about:'
-                ),
-                'confirmation_alert_prompt': expected_alert,
-                'confirmation_alert_prompt_warning': '',
-                'sharable_items_footer': (
-                    'My permission applies only to data from courses or programs that are sponsored by '
-                    'Starfleet Academy, and not to data from any My Platform courses or programs that '
-                    'I take on my own. I understand that once I grant my permission to allow data to be shared '
-                    'with Starfleet Academy, I may not withdraw my permission but I may elect to unenroll '
-                    'from any courses that are sponsored by Starfleet Academy.'
-                ),
-                "course_id": "course-v1:edX+DemoX+Demo_Course",
-                "redirect_url": "https://google.com",
-                "enterprise_customer_name": ecu.enterprise_customer.name,
-                "course_specific": True,
-                "enrollment_deferred": enrollment_deferred,
-                "welcome_text": "Welcome to My Platform.",
-                'sharable_items_note_header': 'Please note',
-        }.items():
-            assert response.context[key] == value  # pylint:disable=no-member
-
-    @mock.patch('enterprise.views.render', side_effect=fake_render)
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.views.CourseApiClient')
-    def test_get_course_specific_consent_invalid_get_params(
-            self,
-            course_api_client_mock,
-            mock_config,
-            *args
-    ):  # pylint: disable=unused-argument
-        course_id = 'course-v1:edX+DemoX+Demo_Course'
-        mock_config.get_value.return_value = 'My Platform'
-        client = course_api_client_mock.return_value
-        client.get_course_details.return_value = {
-            'name': 'edX Demo Course',
-        }
-        self._login()
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-        )
-        ecu = EnterpriseCustomerUserFactory(
-            user_id=self.user.id,
-            enterprise_customer=enterprise_customer
-        )
-        EnterpriseCourseEnrollment.objects.create(
-            enterprise_customer_user=ecu,
-            course_id=course_id
-        )
-        DataSharingConsentFactory(
-            username=self.user.username,
-            course_id=course_id,
-            enterprise_customer=enterprise_customer,
-        )
-        params = {
-            'course_id': 'course-v1:edX+DemoX+Demo_Course',
-            'next': 'https://google.com',
-            'enrollment_deferred': True,
-        }
-        response = self.client.get(self.url, data=params)
-        assert response.status_code == 404
-
-    @mock.patch('enterprise.views.render', side_effect=fake_render)
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.views.CourseApiClient')
-    def test_get_course_specific_consent_unauthenticated_user(
-            self,
-            course_api_client_mock,
-            mock_config,
-            *args
-    ):  # pylint: disable=unused-argument
-        course_id = 'course-v1:edX+DemoX+Demo_Course'
-        mock_config.get_value.return_value = 'My Platform'
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-        )
-        client = course_api_client_mock.return_value
-        client.get_course_details.return_value = {
-            'name': 'edX Demo Course',
-        }
-        ecu = EnterpriseCustomerUserFactory(
-            user_id=self.user.id,
-            enterprise_customer=enterprise_customer
-        )
-        EnterpriseCourseEnrollment.objects.create(
-            enterprise_customer_user=ecu,
-            course_id=course_id
-        )
-        DataSharingConsentFactory(
-            username=self.user.username,
-            course_id=course_id,
-            enterprise_customer=enterprise_customer,
-        )
-        response = self.client.get(
-            self.url + '?course_id=course-v1%3AedX%2BDemoX%2BDemo_Course&next=https%3A%2F%2Fgoogle.com'
-        )
-        assert response.status_code == 302
-        self.assertRedirects(
-            response,
-            (
-                '/accounts/login/?next=/enterprise/grant_data_sharing_permissions%3Fcourse_id%3Dcourse-v1'
-                '%253AedX%252BDemoX%252BDemo_Course%26next%3Dhttps%253A%252F%252Fgoogle.com'
-            ),
-            fetch_redirect_response=False
-        )
-
-    @mock.patch('enterprise.views.render', side_effect=fake_render)
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.views.CourseApiClient')
-    def test_get_course_specific_consent_bad_api_response(
-            self,
-            course_api_client_mock,
-            mock_config,
-            *args
-    ):  # pylint: disable=unused-argument
-        self._login()
-        course_id = 'course-v1:edX+DemoX+Demo_Course'
-        mock_config.get_value.return_value = 'My Platform'
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-        )
-        client = course_api_client_mock.return_value
-        client.get_course_details.side_effect = HttpClientError
-        ecu = EnterpriseCustomerUserFactory(
-            user_id=self.user.id,
-            enterprise_customer=enterprise_customer
-        )
-        EnterpriseCourseEnrollment.objects.create(
-            enterprise_customer_user=ecu,
-            course_id=course_id
-        )
-        DataSharingConsentFactory(
-            username=self.user.username,
-            course_id=course_id,
-            enterprise_customer=enterprise_customer,
-        )
-        response = self.client.get(
-            self.url + '?course_id=course-v1%3AedX%2BDemoX%2BDemo_Course&next=https%3A%2F%2Fgoogle.com'
-        )
-        assert response.status_code == 404
-
-    @mock.patch('enterprise.views.render', side_effect=fake_render)
-    @mock.patch('enterprise.models.CourseCatalogApiServiceClient')
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.views.CourseApiClient')
-    def test_get_course_specific_consent_not_needed(
-            self,
-            course_api_client_mock,
-            mock_config,
-            course_catalog_api_client_mock,
-            *args
-    ):  # pylint: disable=unused-argument
-        self._login()
-        course_id = 'course-v1:edX+DemoX+Demo_Course'
-        mock_config.get_value.return_value = 'My Platform'
-        course_catalog_api_client = course_catalog_api_client_mock.return_value
-        course_catalog_api_client.is_course_in_catalog.return_value = False
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-        )
-        client = course_api_client_mock.return_value
-        client.get_course_details.return_value = {
-            'name': 'edX Demo Course',
-        }
-        ecu = EnterpriseCustomerUserFactory(
-            user_id=self.user.id,
-            enterprise_customer=enterprise_customer
-        )
-        EnterpriseCourseEnrollment.objects.create(
-            enterprise_customer_user=ecu,
-            course_id=course_id,
-            consent_granted=True,
-        )
-        DataSharingConsentFactory(
-            username=self.user.username,
-            course_id=course_id,
-            enterprise_customer=enterprise_customer,
-            granted=True
-        )
-        response = self.client.get(
-            self.url + '?course_id=course-v1%3AedX%2BDemoX%2BDemo_Course&next=https%3A%2F%2Fgoogle.com'
-        )
-        assert response.status_code == 404
-
-    @mock.patch('enterprise.views.render', side_effect=fake_render)
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.views.CourseApiClient')
-    @mock.patch('enterprise.views.reverse')
-    @ddt.data(
-        (True, True, '/successful_enrollment'),
-        (False, True, '/successful_enrollment'),
-        (True, False, '/failure_url'),
-        (False, False, '/failure_url'),
-    )
-    @ddt.unpack
-    def test_post_course_specific_consent(
-            self,
-            enrollment_deferred,
-            consent_provided,
-            expected_redirect_url,
-            reverse_mock,
-            course_api_client_mock,
-            *args
-    ):  # pylint: disable=unused-argument
-        self._login()
-        course_id = 'course-v1:edX+DemoX+Demo_Course'
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-        )
-        ecu = EnterpriseCustomerUserFactory(
-            user_id=self.user.id,
-            enterprise_customer=enterprise_customer
-        )
-        enrollment = EnterpriseCourseEnrollment.objects.create(
-            enterprise_customer_user=ecu,
-            course_id=course_id,
-            consent_granted=consent_provided,
-        )
-        DataSharingConsentFactory(
-            username=self.user.username,
-            course_id=course_id,
-            enterprise_customer=enterprise_customer,
-            granted=consent_provided
-        )
-        client = course_api_client_mock.return_value
-        client.get_course_details.return_value = {
-            'name': 'edX Demo Course',
-        }
-        reverse_mock.return_value = '/dashboard'
-        post_data = {
-            'course_id': course_id,
-            'redirect_url': '/successful_enrollment',
-            'failure_url': '/failure_url',
-        }
-        if enrollment_deferred:
-            post_data['enrollment_deferred'] = True
-        if consent_provided:
-            post_data['data_sharing_consent'] = consent_provided
-
-        resp = self.client.post(self.url, post_data)
-
-        assert resp.url.endswith(expected_redirect_url)  # pylint: disable=no-member
-        assert resp.status_code == 302
-        enrollment.refresh_from_db()
-        if not enrollment_deferred:
-            assert enrollment.consent_granted is consent_provided
-
-    @mock.patch('enterprise.views.render', side_effect=fake_render)
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.views.CourseApiClient')
-    @mock.patch('enterprise.views.reverse')
-    def test_post_course_specific_consent_no_user(
-            self,
-            reverse_mock,
-            course_api_client_mock,
-            *args
-    ):  # pylint: disable=unused-argument
-        course_id = 'course-v1:edX+DemoX+Demo_Course'
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-        )
-        ecu = EnterpriseCustomerUserFactory(
-            user_id=self.user.id,
-            enterprise_customer=enterprise_customer
-        )
-        EnterpriseCourseEnrollment.objects.create(
-            enterprise_customer_user=ecu,
-            course_id=course_id
-        )
-        DataSharingConsentFactory(
-            username=self.user.username,
-            course_id=course_id,
-            enterprise_customer=enterprise_customer,
-        )
-        client = course_api_client_mock.return_value
-        client.get_course_details.return_value = {
-            'name': 'edX Demo Course',
-        }
-        reverse_mock.return_value = '/dashboard'
-        resp = self.client.post(
-            self.url,
-            data={
-                'course_id': course_id,
-                'redirect_url': '/successful_enrollment'
-            },
-        )
-        assert resp.status_code == 302
-        self.assertRedirects(
-            resp,
-            '/accounts/login/?next=/enterprise/grant_data_sharing_permissions',
-            fetch_redirect_response=False
-        )
-
-    @mock.patch('enterprise.views.render', side_effect=fake_render)
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.views.CourseApiClient')
-    @mock.patch('enterprise.views.reverse')
-    def test_post_course_specific_consent_bad_api_response(
-            self,
-            reverse_mock,
-            course_api_client_mock,
-            *args
-    ):  # pylint: disable=unused-argument
-        self._login()
-        course_id = 'course-v1:does+not+exist'
-        data_sharing_consent = True
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-        )
-        ecu = EnterpriseCustomerUserFactory(
-            user_id=self.user.id,
-            enterprise_customer=enterprise_customer
-        )
-        enrollment = EnterpriseCourseEnrollment.objects.create(
-            enterprise_customer_user=ecu,
-            course_id=course_id
-        )
-        DataSharingConsentFactory(
-            username=self.user.username,
-            course_id=course_id,
-            enterprise_customer=enterprise_customer,
-        )
-        client = course_api_client_mock.return_value
-        client.get_course_details.side_effect = HttpClientError
-        reverse_mock.return_value = '/dashboard'
-        resp = self.client.post(
-            self.url,
-            data={
-                'course_id': course_id,
-                'data_sharing_consent': data_sharing_consent,
-                'redirect_url': '/successful_enrollment'
-            },
-        )
-        assert resp.status_code == 404
-        enrollment.refresh_from_db()
-        assert enrollment.consent_granted is None
-
-
-class TestPushLearnerDataToIntegratedChannel(TestCase):
-    """
-    Test PushLearnerDataToIntegratedChannel.
-    """
-
-    url = reverse('push_learner_data')
-
-    def test_post(self):
-        client = Client()
-        try:
-            client.post(self.url)
-            self.fail("Should have raised NotImplementedError")
-        except NotImplementedError:
-            pass
-
-
-class TestPushCatalogDataToIntegratedChannel(TestCase):
-    """
-    Test PushCatalogDataToIntegratedChannel.
-    """
-
-    url = reverse('push_catalog_data')
-
-    def test_post(self):
-        client = Client()
-        try:
-            client.post(self.url)
-            self.fail("Should have raised NotImplementedError")
-        except NotImplementedError:
-            pass
 
 
 @mark.django_db
@@ -1838,289 +1114,3 @@ class TestCourseEnrollmentView(MessagesMixin, TestCase):
         self._login()
         response = self.client.get(enterprise_landing_page_url)
         self._check_expected_enrollment_page(response, expected_context)
-
-
-@mark.django_db
-@ddt.ddt
-class TestHandleConsentEnrollmentView(TestCase):
-    """
-    Test HandleConsentEnrollment.
-    """
-
-    def setUp(self):
-        self.user = UserFactory.create(is_staff=True, is_active=True)
-        self.user.set_password("QWERTY")
-        self.user.save()
-        self.client = Client()
-        self.demo_course_id = 'course-v1:edX+DemoX+Demo_Course'
-        self.dummy_demo_course_modes = [
-            {
-                "slug": "professional",
-                "name": "Professional Track",
-                "min_price": 100,
-                "sku": "sku-audit",
-            },
-            {
-                "slug": "audit",
-                "name": "Audit Track",
-                "min_price": 0,
-                "sku": "sku-audit",
-            },
-        ]
-        super(TestHandleConsentEnrollmentView, self).setUp()
-
-    def _login(self):
-        """
-        Log user in.
-        """
-        assert self.client.login(username=self.user.username, password="QWERTY")
-
-    def _setup_registry_mock(self, registry_mock, provider_id):
-        """
-        Sets up the SSO Registry object
-        """
-        registry_mock.get.return_value.configure_mock(provider_id=provider_id, drop_existing_session=False)
-
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.utils.Registry')
-    def test_handle_consent_enrollment_without_course_mode(
-            self,
-            registry_mock,
-            *args
-    ):  # pylint: disable=unused-argument
-        """
-        Verify that user is redirected to LMS dashboard in case there is
-        no parameter `course_mode` in the request querystring.
-        """
-        course_id = self.demo_course_id
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-            enable_audit_enrollment=True,
-        )
-        faker = FakerFactory.create()
-        provider_id = faker.slug()  # pylint: disable=no-member
-        self._setup_registry_mock(registry_mock, provider_id)
-        EnterpriseCustomerIdentityProviderFactory(provider_id=provider_id, enterprise_customer=enterprise_customer)
-        self._login()
-        handle_consent_enrollment_url = reverse(
-            'enterprise_handle_consent_enrollment',
-            args=[enterprise_customer.uuid, course_id],
-        )
-        response = self.client.get(handle_consent_enrollment_url)
-        redirect_url = LMS_DASHBOARD_URL
-        self.assertRedirects(response, redirect_url, fetch_redirect_response=False)
-
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.views.EnrollmentApiClient')
-    @mock.patch('enterprise.utils.Registry')
-    def test_handle_consent_enrollment_404(
-            self,
-            registry_mock,
-            enrollment_api_client_mock,
-            *args
-    ):  # pylint: disable=unused-argument
-        """
-        Verify that user gets HTTP 404 response if there is no enterprise in
-        database against the provided enterprise UUID or if enrollment API
-        client is unable to get course modes for the provided course id.
-        """
-        course_id = self.demo_course_id
-        enrollment_client = enrollment_api_client_mock.return_value
-        enrollment_client.get_course_modes.side_effect = HttpClientError
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-            enable_audit_enrollment=True,
-        )
-        faker = FakerFactory.create()
-        provider_id = faker.slug()  # pylint: disable=no-member
-        self._setup_registry_mock(registry_mock, provider_id)
-        EnterpriseCustomerIdentityProviderFactory(provider_id=provider_id, enterprise_customer=enterprise_customer)
-        self._login()
-        handle_consent_enrollment_url = '{consent_enrollment_url}?{params}'.format(
-            consent_enrollment_url=reverse(
-                'enterprise_handle_consent_enrollment', args=[enterprise_customer.uuid, course_id]
-            ),
-            params=urlencode({'course_mode': 'professional'})
-        )
-        response = self.client.get(handle_consent_enrollment_url)
-        assert response.status_code == 404
-
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.views.EnrollmentApiClient')
-    @mock.patch('enterprise.utils.Registry')
-    def test_handle_consent_enrollment_no_enterprise_user(
-            self,
-            registry_mock,
-            enrollment_api_client_mock,
-            *args
-    ):  # pylint: disable=unused-argument
-        """
-        Verify that user gets HTTP 404 response if the user is not linked to
-        the enterprise with the provided enterprise UUID or if enrollment API
-        client is unable to get course modes for the provided course id.
-        """
-        course_id = self.demo_course_id
-        enrollment_client = enrollment_api_client_mock.return_value
-        enrollment_client.get_course_modes.return_value = self.dummy_demo_course_modes
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-            enable_audit_enrollment=True,
-        )
-        faker = FakerFactory.create()
-        provider_id = faker.slug()  # pylint: disable=no-member
-        self._setup_registry_mock(registry_mock, provider_id)
-        EnterpriseCustomerIdentityProviderFactory(provider_id=provider_id, enterprise_customer=enterprise_customer)
-        self._login()
-        handle_consent_enrollment_url = '{consent_enrollment_url}?{params}'.format(
-            consent_enrollment_url=reverse(
-                'enterprise_handle_consent_enrollment', args=[enterprise_customer.uuid, course_id]
-            ),
-            params=urlencode({'course_mode': 'professional'})
-        )
-        response = self.client.get(handle_consent_enrollment_url)
-        assert response.status_code == 404
-
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.views.get_enterprise_customer_user')
-    @mock.patch('enterprise.views.EnrollmentApiClient')
-    @mock.patch('enterprise.utils.Registry')
-    def test_handle_consent_enrollment_with_invalid_course_mode(
-            self,
-            registry_mock,
-            enrollment_api_client_mock,
-            get_ec_user_mock,
-            *args
-    ):  # pylint: disable=unused-argument
-        """
-        Verify that user is redirected to LMS dashboard in case the provided
-        course mode does not exist.
-        """
-        course_id = self.demo_course_id
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-            enable_audit_enrollment=True,
-        )
-        faker = FakerFactory.create()
-        provider_id = faker.slug()  # pylint: disable=no-member
-        self._setup_registry_mock(registry_mock, provider_id)
-        EnterpriseCustomerIdentityProviderFactory(provider_id=provider_id, enterprise_customer=enterprise_customer)
-        enterprise_customer_user = EnterpriseCustomerUserFactory(
-            user_id=self.user.id,
-            enterprise_customer=enterprise_customer
-        )
-        enrollment_client = enrollment_api_client_mock.return_value
-        enrollment_client.get_course_modes.return_value = self.dummy_demo_course_modes
-        mocked_enterprise_customer_user = get_ec_user_mock.return_value
-        mocked_enterprise_customer_user.return_value = enterprise_customer_user
-        self._login()
-        handle_consent_enrollment_url = '{consent_enrollment_url}?{params}'.format(
-            consent_enrollment_url=reverse(
-                'enterprise_handle_consent_enrollment', args=[enterprise_customer.uuid, course_id]
-            ),
-            params=urlencode({'course_mode': 'some-invalid-course-mode'})
-        )
-        response = self.client.get(handle_consent_enrollment_url)
-        redirect_url = LMS_DASHBOARD_URL
-        self.assertRedirects(response, redirect_url, fetch_redirect_response=False)
-
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.views.EnrollmentApiClient')
-    @mock.patch('enterprise.utils.Registry')
-    def test_handle_consent_enrollment_with_audit_course_mode(
-            self,
-            registry_mock,
-            enrollment_api_client_mock,
-            *args
-    ):  # pylint: disable=unused-argument
-        """
-        Verify that user is redirected to course in case the provided
-        course mode is audit track.
-        """
-        course_id = self.demo_course_id
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-            enable_audit_enrollment=True,
-        )
-        faker = FakerFactory.create()
-        provider_id = faker.slug()  # pylint: disable=no-member
-        self._setup_registry_mock(registry_mock, provider_id)
-        EnterpriseCustomerIdentityProviderFactory(provider_id=provider_id, enterprise_customer=enterprise_customer)
-        enterprise_customer_user = EnterpriseCustomerUserFactory(
-            user_id=self.user.id,
-            enterprise_customer=enterprise_customer
-        )
-        enrollment_client = enrollment_api_client_mock.return_value
-        enrollment_client.get_course_modes.return_value = self.dummy_demo_course_modes
-        self._login()
-        handle_consent_enrollment_url = '{consent_enrollment_url}?{params}'.format(
-            consent_enrollment_url=reverse(
-                'enterprise_handle_consent_enrollment', args=[enterprise_customer.uuid, course_id]
-            ),
-            params=urlencode({'course_mode': 'audit'})
-        )
-        response = self.client.get(handle_consent_enrollment_url)
-        redirect_url = LMS_COURSEWARE_URL.format(course_id=course_id)
-        self.assertRedirects(response, redirect_url, fetch_redirect_response=False)
-
-        self.assertTrue(EnterpriseCourseEnrollment.objects.filter(
-            enterprise_customer_user__enterprise_customer=enterprise_customer,
-            enterprise_customer_user__user_id=enterprise_customer_user.user_id,
-            course_id=course_id
-        ).exists())
-
-    @mock.patch('enterprise.views.configuration_helpers')
-    @mock.patch('enterprise.views.EnrollmentApiClient')
-    @mock.patch('enterprise.utils.Registry')
-    def test_handle_consent_enrollment_with_professional_course_mode(
-            self,
-            registry_mock,
-            enrollment_api_client_mock,
-            *args
-    ):  # pylint: disable=unused-argument
-        """
-        Verify that user is redirected to course in case the provided
-        course mode is audit track.
-        """
-        course_id = self.demo_course_id
-        enterprise_customer = EnterpriseCustomerFactory(
-            name='Starfleet Academy',
-            enable_data_sharing_consent=True,
-            enforce_data_sharing_consent='at_enrollment',
-            enable_audit_enrollment=True,
-        )
-        faker = FakerFactory.create()
-        provider_id = faker.slug()  # pylint: disable=no-member
-        self._setup_registry_mock(registry_mock, provider_id)
-        EnterpriseCustomerIdentityProviderFactory(provider_id=provider_id, enterprise_customer=enterprise_customer)
-        enterprise_customer_user = EnterpriseCustomerUserFactory(
-            user_id=self.user.id,
-            enterprise_customer=enterprise_customer
-        )
-        enrollment_client = enrollment_api_client_mock.return_value
-        enrollment_client.get_course_modes.return_value = self.dummy_demo_course_modes
-        self._login()
-        handle_consent_enrollment_url = '{consent_enrollment_url}?{params}'.format(
-            consent_enrollment_url=reverse(
-                'enterprise_handle_consent_enrollment', args=[enterprise_customer.uuid, course_id]
-            ),
-            params=urlencode({'course_mode': 'professional'})
-        )
-        response = self.client.get(handle_consent_enrollment_url)
-        redirect_url = LMS_START_PREMIUM_COURSE_FLOW_URL.format(course_id=course_id)
-        self.assertRedirects(response, redirect_url, fetch_redirect_response=False)
-
-        self.assertTrue(EnterpriseCourseEnrollment.objects.filter(
-            enterprise_customer_user__enterprise_customer=enterprise_customer,
-            enterprise_customer_user__user_id=enterprise_customer_user.user_id,
-            course_id=course_id
-        ).exists())
